ДобавCRUD запроÑсы для matches

+++ src/main/java/com/example/snake/contest/dto/CreatePlayerDTO.java
@@ -5,14 +5,14 @@ import jakarta.validation.constraints.NotNull;
 public class CreatePlayerDTO {
 
     @NotNull
-    private Long participant_id;
+    private Integer participant_id;
 
 //    Сеттеры геттеры
-    public @NotNull Long getParticipant_id() {
+    public @NotNull Integer getParticipant_id() {
         return participant_id;
     }
 
-    public void setParticipant_id(@NotNull Long participant_id) {
+    public void setParticipant_id(@NotNull Integer participant_id) {
         this.participant_id = participant_id;
     }
